requirements.txt
streamlit>=1.31.0
supabase>=2.7.0
python-dotenv>=1.0.0
streamlit-supabase-auth
pillow>=10.0.0
requests>=2.31.0